Fixed formatting issues + switch from styled components to css
bootstrap uncommented from index.js, fixed formatting with bootstrap and original text formatting / calendar
