+++ src/Components/HomePage/HomeNav.css
@@ -1,0 +1,38 @@
+.homepage-divider {
+  height: 3px;
+  background-color: rgb(1, 114, 71);
+  border-style: none;
+  margin-top: 40px;
+  position: fixed;
+  width: 100%;
+  z-index: 1;
+}
+
+.homepage-nav-bar {
+  padding-top: 90px;
+  padding-bottom: 15px;
+  text-align: center;
+  background-color: white;
+  font-size: 15px;
+  cursor: pointer;
+  font-weight: bold;
+}
+
+.homepage-nav-bar-button {
+
+}
+
+.homepage-nav-bar-button:hover {
+  color: rgb(1, 114, 71);
+  background-color: white;  
+}
+
+.homepage-tab-container {
+  display: flex;
+  justify-content: space-around;
+  position: fixed;
+  width: 100%;
+  background-color: white;
+  z-index: 1;
+  height: 55px;
+}
